refactor: Update GizmoManager import and fix alignment method calls
