+++ packages/@dcl/inspector/src/lib/babylon/decentraland/SceneContext.ts
@@ -15,7 +15,6 @@ import { putTransformComponent } from './sdkComponents/transform'
 import { putTextShapeComponent } from './sdkComponents/text-shape'
 import { putSceneComponent } from './editorComponents/scene'
 import { createOperations } from '../../sdk/operations'
-// import { createGizmoManager } from './gizmo-manager'
 import { createGizmoManager } from './GizmoManager'
 import { getDataLayerInterface } from '../../../redux/data-layer'
 import { putMaterialComponent } from './sdkComponents/material'
