+++ packages/@dcl/inspector/src/lib/sdk/context.ts
@@ -9,7 +9,7 @@ import { EditorComponents, SdkComponents } from './components'
 import { getHardcodedLoadableScene } from './test-local-scene'
 import { createInspectorEngine } from './inspector-engine'
 import { createOperations } from './operations'
-import { Gizmos } from '../babylon/decentraland/gizmo-manager'
+import { Gizmos } from '../babylon/decentraland/GizmoManager'
 import { CameraManager } from '../babylon/decentraland/camera'
 import { InspectorPreferences } from '../logic/preferences/types'
 import { CameraServer } from '../rpc/camera/server'
